+++ templates/general.js
@@ -22,6 +22,17 @@ ${objectClassName};
 }, {
   name: () => 'package.json',
   content: ({ moduleName, platforms, githubAccount, authorName, authorEmail, license }) => {
+    const files =
+      `[
+    "README.md",` +
+    (platforms.indexOf('android') >= 0 ? `
+    "android",` : ``) + `
+    "index.js"` +
+    (platforms.indexOf('ios') >= 0 ? `,
+    "ios",
+    "${moduleName}.podspec"` : ``) + `
+  ]`;
+
     const peerDependencies =
       `{
     "react": "^16.8.1",
@@ -40,6 +51,7 @@ ${objectClassName};
   "version": "1.0.0",
   "description": "TODO",
   "main": "index.js",
+  "files": ${files},
   "scripts": {
     "test": "echo \\"Error: no test specified\\" && exit 1"
   },
@@ -62,7 +74,7 @@ ${objectClassName};
   "devDependencies": ${devDependencies}
 }
 `;
-  },
+  }
 }, {
   // for module without view:
   name: ({ view }) => !view && 'index.js',
@@ -141,23 +153,5 @@ buck-out/
     }
 
     return content;
-  },
-}, {
-  name: () => '.gitattributes',
-  content: ({ platforms }) => {
-    if (platforms.indexOf('ios') >= 0) {
-      return '*.pbxproj -text\n';
-    }
-
-    return '';
-  }
-}, {
-  name: () => '.npmignore',
-  content: ({ generateExample, exampleName }) => {
-    if (generateExample) {
-      return `${exampleName}\n`;
-    }
-
-    return '';
   }
 }];