+++ tests/unsupported-platforms/windows-csharp/with-options/windows-csharp-with-options.test.js
@@ -6,7 +6,7 @@ const uuid = 'E22606E0-B47F-11E9-A3F0-07F70A25DAFB';
 test('unsupported Windows C# template with options', () => {
   const options = {
     // normalized for name: 'alice-bobbi':
-    name: 'AliceBobbi',
+    objectClassName: 'AliceBobbi',
     // options
     githubAccount: 'alicebits',
     authorName: 'Alice',